[IMP] mail: move member list code to discuss
Some getters and fields related to `Member List` are defined
in `core/common` while they are used in discuss only.
So keeping them in `core` is useless.

This commit moves this code to `discuss` folder.

Also `sortOnlineMembers` has been moved to `Store`, so that this
method is no longer static. New static methods cannot be augmented
in models interface, so by it being initially defined in a patch
would lead to lack of typing. This is undesirable for reducing
programming errors.
Since the context of this method is not coupled to an instance of
Thread, this has been moved to Store which is readily available
for general-purpose methods that could be overridden.
