+++ addons/mail/static/src/discuss/core/common/@types/models.d.ts
@@ -1,0 +1,11 @@
+declare module "models" {
+    export interface Store {
+        sortOnlineMembers(m1: ChannelMember, m2: ChannelMember)
+    }
+
+    export interface Thread {
+        onlineMembers: ChannelMember[],
+        offlineMembers: ChannelMember[],
+        readonly hasMemberList: boolean,
+    }
+}
